+++ content/abilities/command.gd
@@ -1,0 +1,11 @@
+class_name Command
+
+var unit: Unit
+var target: Vector2i
+
+func _init(_unit, _target) -> void:
+	unit = _unit
+	target = _target
+
+func execute():
+	pass